re-organize API data structures
