+++ bin/opteadm/src/lib.rs
@@ -18,6 +18,10 @@ use oxide_vpc::api::ClearVirt2BoundaryReq;
 use oxide_vpc::api::CreateXdeReq;
 use oxide_vpc::api::DeleteXdeReq;
 use oxide_vpc::api::DhcpCfg;
+use oxide_vpc::api::DumpVirt2BoundaryReq;
+use oxide_vpc::api::DumpVirt2BoundaryResp;
+use oxide_vpc::api::DumpVirt2PhysReq;
+use oxide_vpc::api::DumpVirt2PhysResp;
 use oxide_vpc::api::FirewallRule;
 use oxide_vpc::api::ListPortsResp;
 use oxide_vpc::api::RemFwRuleReq;
@@ -26,7 +30,6 @@ use oxide_vpc::api::SetFwRulesReq;
 use oxide_vpc::api::SetVirt2BoundaryReq;
 use oxide_vpc::api::SetVirt2PhysReq;
 use oxide_vpc::api::VpcCfg;
-use oxide_vpc::engine::overlay;
 use std::fs::File;
 use std::fs::OpenOptions;
 use std::os::unix::io::AsRawFd;
@@ -219,12 +222,12 @@ impl OpteAdm {
     }
 
     /// Dump the Virtual-to-Physical mappings.
-    pub fn dump_v2p(&self) -> Result<overlay::DumpVirt2PhysResp, Error> {
+    pub fn dump_v2p(&self) -> Result<DumpVirt2PhysResp, Error> {
         let cmd = OpteCmd::DumpVirt2Phys;
         run_cmd_ioctl(
             self.device.as_raw_fd(),
             cmd,
-            Some(&overlay::DumpVirt2PhysReq { unused: 99 }),
+            Some(&DumpVirt2PhysReq { unused: 99 }),
         )
     }
 
@@ -242,12 +245,12 @@ impl OpteAdm {
     }
 
     /// Dump the Virtual-to-Boundary mappings.
-    pub fn dump_v2b(&self) -> Result<overlay::DumpVirt2BoundaryResp, Error> {
+    pub fn dump_v2b(&self) -> Result<DumpVirt2BoundaryResp, Error> {
         let cmd = OpteCmd::DumpVirt2Boundary;
         run_cmd_ioctl(
             self.device.as_raw_fd(),
             cmd,
-            Some(&overlay::DumpVirt2BoundaryReq { unused: 99 }),
+            Some(&DumpVirt2BoundaryReq { unused: 99 }),
         )
     }
