+++ lib/opte-ioctl/src/lib.rs
@@ -17,13 +17,13 @@ use oxide_vpc::api::ClearVirt2BoundaryReq;
 use oxide_vpc::api::CreateXdeReq;
 use oxide_vpc::api::DeleteXdeReq;
 use oxide_vpc::api::DhcpCfg;
+use oxide_vpc::api::DumpVirt2PhysResp;
 use oxide_vpc::api::ListPortsResp;
 use oxide_vpc::api::SetExternalIpsReq;
 use oxide_vpc::api::SetFwRulesReq;
 use oxide_vpc::api::SetVirt2BoundaryReq;
 use oxide_vpc::api::SetVirt2PhysReq;
 use oxide_vpc::api::VpcCfg;
-use oxide_vpc::engine::overlay::DumpVirt2PhysResp;
 use serde::de::DeserializeOwned;
 use serde::Serialize;
 use std::fs::File;
@@ -143,7 +143,7 @@ impl OpteHdl {
         })
     }
 
-    pub fn list_v2p(&self) -> Result<DumpVirt2PhysResp, Error> {
+    pub fn dump_v2p(&self) -> Result<DumpVirt2PhysResp, Error> {
         let cmd = OpteCmd::DumpVirt2Phys;
         run_cmd_ioctl(self.device.as_raw_fd(), cmd, None::<&()>)
     }
